# Fix: Improves performance (refs #3530)