--- a/lib/rules/indent.js
+++ b/lib/rules/indent.js
@@ -75,6 +75,11 @@ module.exports = function(context) {
         }
     }
 
+    var indentPattern = {
+        normal: indentType === "space" ? /^ +/ : /^\t+/,
+        excludeCommas: indentType === "space" ? /^[ ,]+/ : /^[\t,]+/
+    };
+
     var caseIndentStore = {};
 
     /**
@@ -168,17 +173,9 @@ module.exports = function(context) {
     function getNodeIndent(node, byLastLine, excludeCommas) {
         var token = byLastLine ? context.getLastToken(node) : context.getFirstToken(node);
         var src = context.getSource(token, token.loc.start.column);
-
-        var skip = excludeCommas ? "," : "";
-
-        var regExp;
-        if (indentType === "space") {
-            regExp = new RegExp("^[ " + skip + "]+");
-        } else {
-            regExp = new RegExp("^[\t" + skip + "]+");
-        }
-
+        var regExp = excludeCommas ? indentPattern.excludeCommas : indentPattern.normal;
         var indent = regExp.exec(src);
+
         return indent ? indent[0].length : 0;
     }