--- a/lib/rules/no-alert.js
+++ b/lib/rules/no-alert.js
@@ -69,9 +69,8 @@ function findReference(scope, node) {
  * @returns {boolean} Whether or not the name is shadowed globally.
  */
 function isGloballyShadowed(globalScope, identifierName) {
-    return globalScope.variables.some(function(variable) {
-        return variable.name === identifierName && variable.defs.length > 0;
-    });
+    var variable = globalScope.set.get(identifierName);
+    return Boolean(variable && variable.defs.length > 0);
 }
 
 /**