--- a/lib/rules/no-shadow.js
+++ b/lib/rules/no-shadow.js
@@ -7,6 +7,12 @@
 "use strict";
 
 //------------------------------------------------------------------------------
+// Requirements
+//------------------------------------------------------------------------------
+
+var astUtils = require("../ast-utils");
+
+//------------------------------------------------------------------------------
 // Rule Definition
 //------------------------------------------------------------------------------
 
@@ -100,69 +106,36 @@ module.exports = function(context) {
     }
 
     /**
-     * Checks if a variable is contained in the list of given scope variables.
-     * @param {Object} variable The variable to check.
-     * @param {Array} scopeVars The scope variables to look for.
-     * @returns {boolean} Whether or not the variable is contains in the list of scope variables.
-     */
-    function isContainedInScopeVars(variable, scopeVars) {
-        return scopeVars.some(function(scopeVar) {
-            return (
-                (scopeVar.identifiers.length > 0 || (options.builtinGlobals && "writeable" in scopeVar)) &&
-                variable.name === scopeVar.name &&
-                !isDuplicatedClassNameVariable(scopeVar) &&
-                !isOnInitializer(variable, scopeVar) &&
-                !(options.hoist !== "all" && isInTdz(variable, scopeVar))
-            );
-        });
-    }
-
-    /**
-     * Checks if the given variables are shadowed in the given scope.
-     * @param {Array} variables The variables to look for
-     * @param {Object} scope The scope to be checked.
-     * @returns {Array} Variables which are not declared in the given scope.
-     */
-    function checkShadowsInScope(variables, scope) {
-
-        var passedVars = [];
-
-        variables.forEach(function(variable) {
-            // "arguments" is a special case that has no identifiers (#1759)
-            if (variable.identifiers.length > 0 && isContainedInScopeVars(variable, scope.variables)) {
-                context.report(
-                    variable.identifiers[0],
-                    "\"{{name}}\" is already declared in the upper scope.",
-                    {name: variable.name});
-            } else {
-                passedVars.push(variable);
-            }
-        });
-
-        return passedVars;
-    }
-
-    /**
      * Checks the current context for shadowed variables.
      * @param {Scope} scope - Fixme
      * @returns {void}
      */
     function checkForShadows(scope) {
-        var variables = scope.variables.filter(function(variable) {
-            return (
-                // Skip "arguments".
-                variable.identifiers.length > 0 &&
-                // Skip variables of a class name in the class scope of ClassDeclaration.
-                !isDuplicatedClassNameVariable(variable) &&
-                !isAllowed(variable)
-            );
-        });
+        var variables = scope.variables;
+        for (var i = 0; i < variables.length; ++i) {
+            var variable = variables[i];
 
-        // iterate through the array of variables and find duplicates with the upper scope
-        var upper = scope.upper;
-        while (upper && variables.length) {
-            variables = checkShadowsInScope(variables, upper);
-            upper = upper.upper;
+            // Skips "arguments" or variables of a class name in the class scope of ClassDeclaration.
+            if (variable.identifiers.length === 0 ||
+                isDuplicatedClassNameVariable(variable) ||
+                isAllowed(variable)
+            ) {
+                continue;
+            }
+
+            // Gets shadowed variable.
+            var shadowed = astUtils.getVariableByName(scope.upper, variable.name);
+            if (shadowed &&
+                (shadowed.identifiers.length > 0 || (options.builtinGlobals && "writeable" in shadowed)) &&
+                !isOnInitializer(variable, shadowed) &&
+                !(options.hoist !== "all" && isInTdz(variable, shadowed))
+            ) {
+                context.report({
+                    node: variable.identifiers[0],
+                    message: "\"{{name}}\" is already declared in the upper scope.",
+                    data: variable
+                });
+            }
         }
     }