--- a/lib/rules/no-spaced-func.js
+++ b/lib/rules/no-spaced-func.js
@@ -21,26 +21,26 @@ module.exports = function(context) {
      */
     function detectOpenSpaces(node) {
         var lastCalleeToken = sourceCode.getLastToken(node.callee),
-            tokens = sourceCode.getTokens(node),
-            i = tokens.indexOf(lastCalleeToken),
-            l = tokens.length;
+            prevToken = lastCalleeToken,
+            parenToken = sourceCode.getTokenAfter(lastCalleeToken);
 
-        while (i < l && tokens[i].value !== "(") {
-            ++i;
-        }
-
-        if (i >= l) {
+        if (sourceCode.getLastToken(node).value !== ")") {
             return;
         }
 
+        while (parenToken.value !== "(") {
+            prevToken = parenToken;
+            parenToken = sourceCode.getTokenAfter(parenToken);
+        }
+
         // look for a space between the callee and the open paren
-        if (sourceCode.isSpaceBetweenTokens(tokens[i - 1], tokens[i])) {
+        if (sourceCode.isSpaceBetweenTokens(prevToken, parenToken)) {
             context.report({
                 node: node,
                 loc: lastCalleeToken.loc.start,
                 message: "Unexpected space between function name and paren.",
                 fix: function(fixer) {
-                    return fixer.removeRange([tokens[i - 1].range[1], tokens[i].range[0]]);
+                    return fixer.removeRange([prevToken.range[1], parenToken.range[0]]);
                 }
             });
         }