--- a/lib/rules/no-undef.js
+++ b/lib/rules/no-undef.js
@@ -10,11 +10,13 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-// none!
+var astUtils = require("../ast-utils");
 
 //------------------------------------------------------------------------------
 // Helpers
 //------------------------------------------------------------------------------
+
+var hasOwnProperty = Object.prototype.hasOwnProperty;
 
 /**
  * Check if a variable is an implicit declaration
@@ -35,18 +37,13 @@ function isImplicitGlobal(variable) {
  * @returns {Variable} The variable, or null if ref refers to an undeclared variable.
  */
 function getDeclaredGlobalVariable(scope, ref) {
-    var declaredGlobal = null;
-    scope.variables.some(function(variable) {
-        if (variable.name === ref.identifier.name) {
-            // If it's an implicit global, it must have a `writeable` field (indicating it was declared)
-            if (!isImplicitGlobal(variable) || {}.hasOwnProperty.call(variable, "writeable")) {
-                declaredGlobal = variable;
-                return true;
-            }
-        }
-        return false;
-    });
-    return declaredGlobal;
+    var variable = astUtils.getVariableByName(scope, ref.identifier.name);
+
+    // If it's an implicit global, it must have a `writeable` field (indicating it was declared)
+    if (variable && (!isImplicitGlobal(variable) || hasOwnProperty.call(variable, "writeable"))) {
+        return variable;
+    }
+    return null;
 }
 
 /**