--- a/lib/rules/no-use-before-define.js
+++ b/lib/rules/no-use-before-define.js
@@ -7,6 +7,12 @@
 "use strict";
 
 //------------------------------------------------------------------------------
+// Requirements
+//------------------------------------------------------------------------------
+
+var astUtils = require("../ast-utils");
+
+//------------------------------------------------------------------------------
 // Constants
 //------------------------------------------------------------------------------
 
@@ -17,26 +23,6 @@ var NO_FUNC = "nofunc";
 //------------------------------------------------------------------------------
 
 module.exports = function(context) {
-
-    /**
-     * Finds variable declarations in a given scope.
-     * @param {string} name The variable name to find.
-     * @param {Scope} scope The scope to search in.
-     * @returns {Object} The variable declaration object.
-     * @private
-     */
-    function findDeclaration(name, scope) {
-        // try searching in the current scope first
-        for (var i = 0, l = scope.variables.length; i < l; i++) {
-            if (scope.variables[i].name === name) {
-                return scope.variables[i];
-            }
-        }
-        // check if there's upper scope and call recursivly till we find the variable
-        if (scope.upper) {
-            return findDeclaration(name, scope.upper);
-        }
-    }
 
     /**
      * Finds and validates all variables in a given scope.
@@ -68,7 +54,7 @@ module.exports = function(context) {
             if (reference.resolved && reference.resolved.identifiers.length > 0) {
                 checkLocationAndReport(reference, reference.resolved);
             } else {
-                var declaration = findDeclaration(reference.identifier.name, scope);
+                var declaration = astUtils.getVariableByName(scope, reference.identifier.name);
                 // if there're no identifiers, this is a global environment variable
                 if (declaration && declaration.identifiers.length !== 0) {
                     checkLocationAndReport(reference, declaration);