--- a/lib/rules/space-before-function-paren.js
+++ b/lib/rules/space-before-function-paren.js
@@ -37,7 +37,7 @@ module.exports = function(context) {
             return true;
         }
 
-        parent = context.getAncestors().pop();
+        parent = node.parent;
         return parent.type === "MethodDefinition" ||
             (parent.type === "Property" &&
                 (
@@ -55,7 +55,6 @@ module.exports = function(context) {
      */
     function validateSpacingBeforeParentheses(node) {
         var isNamed = isNamedFunction(node),
-            tokens,
             leftToken,
             rightToken,
             location;
@@ -64,31 +63,11 @@ module.exports = function(context) {
             return;
         }
 
-        tokens = context.getTokens(node);
-
-        if (node.generator) {
-            if (node.id) {
-                leftToken = tokens[2];
-                rightToken = tokens[3];
-            } else {
-                // Object methods are named but don't have an id
-                leftToken = context.getTokenBefore(node);
-                rightToken = tokens[0];
-            }
-        } else if (isNamed) {
-            if (node.id) {
-                leftToken = tokens[1];
-                rightToken = tokens[2];
-            } else {
-                // Object methods are named but don't have an id
-                leftToken = context.getTokenBefore(node);
-                rightToken = tokens[0];
-            }
-        } else {
-            leftToken = tokens[0];
-            rightToken = tokens[1];
+        rightToken = sourceCode.getFirstToken(node);
+        while (rightToken.value !== "(") {
+            rightToken = sourceCode.getTokenAfter(rightToken);
         }
-
+        leftToken = context.getTokenBefore(rightToken);
         location = leftToken.loc.end;
 
         if (sourceCode.isSpaceBetweenTokens(leftToken, rightToken)) {