--- a/lib/rules/spaced-comment.js
+++ b/lib/rules/spaced-comment.js
@@ -41,9 +41,7 @@ function escapeAndRepeat(s) {
  * @returns {string[]} A marker list.
  */
 function parseMarkersOption(markers) {
-    if (!markers) {
-        markers = [];
-    }
+    markers = markers ? markers.slice(0) : [];
 
     // `*` is a marker for JSDoc comments.
     if (markers.indexOf("*") === -1) {