… lib/testers/rule-tester.js