This was separated from #3559 

- `RuleContext` method definitions are moved to outside of the constructor. `RuleContext` object is created many times, so that constructor should be as simple as possible. (`lib/rule-context.js`)
- Construction of a regular expression object which depends on rule's options is moved to outside of `getNodeIndent` function. The function is called many times and to create dynamic regular expression object is slow. (`lib/rules/indent.js`)
- Avoid calling `Function#apply` as possible. (`lib/eslint.js` and `lib/rule-context.js`)
- Avoid calling `SourceCode#getTokens`. A CPU Profiler said that function is slow a bit. (`lib/rules/no-spaced-func.js` and `lib/rules/space-before-function-paren.js`)

**Before** (`npm run perf` for multiple files)

```
CPU Speed is 3392 with multiplier 7500000
Performance Run #1:  3882.190459ms
Performance Run #2:  3767.144016ms
Performance Run #3:  3813.567308ms
Performance Run #4:  3784.897383ms
Performance Run #5:  3768.8099039999997ms
Performance budget exceeded: 3784.897383ms (limit: 2211.0849056603774ms)


Load performance Run #1:  252.483815ms
Load performance Run #2:  251.585219ms
Load performance Run #3:  248.925052ms
Load performance Run #4:  252.786566ms
Load performance Run #5:  253.870494ms

Load Performance median :  252.483815ms
```

**After**

```
CPU Speed is 3392 with multiplier 7500000
Performance Run #1:  3075.235085ms
Performance Run #2:  3030.254616ms
Performance Run #3:  3057.690596ms
Performance Run #4:  3047.723645ms
Performance Run #5:  3016.89371ms
Performance budget exceeded: 3047.723645ms (limit: 2211.0849056603774ms)


Load performance Run #1:  254.633259ms
Load performance Run #2:  250.226913ms
Load performance Run #3:  253.68184ms
Load performance Run #4:  252.021688ms
Load performance Run #5:  251.831827ms

Load Performance median :  252.021688ms
```