@mysticatea 2015-11-15
I updated with more optimizations.

- To search variables by name exists in many places. And currently the search is linear search. Especially, browser environment has many many variables. So I replaced those with O(1) search.
- `RuleTester` comes to check whether or not `RuleContext` instance was not modified by rules.

As a result, in my PC:

|                                          | before | after  | %    |
|------------------------------------------|-------:|-------:|-----:|
| `npm test` time                          | 52s    | 34s    | -35% |
| `npm run perf` (single)                  | 2204ms | 2182ms |  -1% |
| `npm run perf` (multi)                   | 3804ms | 3013ms | -21% |
| `npm run perf` (single with browser env) | 2266ms | 2182ms |  -4% |
| `npm run perf` (multi with browser env)  | 5450ms | 3634ms | -33% |

- single: `--no-ignore ./tests/performance/jshint.js`
- multi: `--no-ignore lib`
- single with browser env: `--no-ignore ./tests/performance/jshint.js --env browser`
- multi with browser env: `--no-ignore lib --env browser`