@mysticatea 2015-11-17
I updated with `Object.freeze(this)` and `(a, b, c, d, e)`!